# feat: compute rank sum via histogram